modification chaine de responsabilite, cast en int

+++ experts/ExpertDessinCercle.java
@@ -31,9 +31,9 @@ public class ExpertDessinCercle extends ExpertDessin
 		{   
 			fen.changerCouleur(tabStrings[1]) ;
 			fen.dessinerCercle(
-					Integer.parseInt(tabStrings[2]),
-					Integer.parseInt(tabStrings[3]),
-					Integer.parseInt(tabStrings[4])
+					(int)Double.parseDouble(tabStrings[2]),
+					(int)Double.parseDouble(tabStrings[3]),
+					(int)Double.parseDouble(tabStrings[4])
 			);
 			System.out.println("c'est un cercle !");
 		}
